move files steps to file context
grammar fix
