+++ test/gui/shared/steps/spaces_context.py
@@ -3,7 +3,6 @@ from pageObjects.EnterPassword import EnterPassword
 from helpers.UserHelper import getDisplaynameForUser, getPasswordForUser
 from helpers.SetupClientHelper import setUpClient, getResourcePath
 from helpers.SyncHelper import waitForInitialSyncToComplete
-from helpers.FilesHelper import can_read, can_write, read_file_content
 from helpers.SpaceHelper import (
     create_space,
     create_space_folder,
@@ -46,27 +45,8 @@ def step(context, user, space_name):
     waitForInitialSyncToComplete(getResourcePath('/', user, space_name))
 
 
-@Then('user "|any|" should be able to open the file "|any|" on the file system')
-def step(context, user, file_name):
-    file_path = getResourcePath(file_name, user)
-    test.compare(can_read(file_path), True, "File should be readable")
-
-
-@Then('as "|any|" the file "|any|" should have content "|any|" on the file system')
-def step(context, user, file_name, content):
-    file_path = getResourcePath(file_name, user)
-    file_content = read_file_content(file_path)
-    test.compare(file_content, content, "Comparing file content")
-
-
-@Then('user "|any|" should not be able to edit the file "|any|" on the file system')
-def step(context, user, file_name):
-    file_path = getResourcePath(file_name, user)
-    test.compare(not can_write(file_path), True, "File should not be writable")
-
-
 @Then(
-    'as "|any|" the file "|any|" of space "|any|" should have content "|any|" in the server'
+    'as "|any|" the file "|any|" in the space "|any|" should have content "|any|" in the server'
 )
 def step(context, user, file_name, space_name, content):
     downloaded_content = get_file_content(space_name, file_name, user)
@@ -80,4 +60,4 @@ def step(context, user, file_name, space_name, content):
 def step(context, user, space_name, resource_name):
     print(user, space_name, resource_name)
     exists = resource_exists(space_name, resource_name, user)
-    test.compare(exists, True, "Resource should exist")
+    test.compare(exists, True, "Resource exists")
